separate legacySession from sessions
