+++ packages/server/src/app.ts
@@ -11,10 +11,11 @@ import { createCapabilitiesRouter } from './apis/capabilities';
 import { createLegacyMcpRouter } from './apis/legacyMcp';
 import { createLegacyMcpOAuthRouter } from './apis/legacyMcpOAuth';
 import { createLegacyModelsRouter } from './apis/legacyModels';
+import { createLegacySessionsRouter } from './apis/legacySessions';
 import { createLegacySkillsRouter } from './apis/legacySkills';
 import { createAvailableMcpServersRouter } from './apis/mcpServers';
 import { createModelsRouter } from './apis/models';
-import { createLegacySessionsRouter, createSessionsRouter } from './apis/sessions';
+import { createSessionsRouter } from './apis/sessions';
 import { createSettingsRouter } from './apis/settings';
 import { createAvailableSkillsRouter } from './apis/skills';
 import { createLegacyTurnsRouter, createTurnsRouter } from './apis/turns';
